Finer control for using mpi-wrappers for compilers
For CUDA builds different codes have different requirements on whether they want
to use MPI compiler wrappers (i.e., mpic++) vs. the compiler executable. This
commit introduces separate variables to allow the codes finer control of
choosing the appropriate path.

For NREL Eagle and ORNL Summit, this reverts the meaning of variables like CXX
to be the compiler and not the MPI wrapper. The MPI wrappers are now available
through MPICXX, MPICC, and MPIFC.